# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sellify API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [me]
      summary: Get current user's profile (creates one on first call)
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateMe
      tags: [me]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
...
  /me/delete:
    post:
      operationId: deleteMyAccount
      tags: [me]
      summary: Permanently delete the current user's account and all associated data
      responses:
        "204":
          description: Account deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /me/push-token:
    post:
      operationId: savePushToken
      tags: [me]
      summary: Register a device push token for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushTokenInput"
      responses:
        "204":
          description: Saved
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
...
    Profile:
      type: object
      required: [id, displayName, memberSince, isModerator]
      properties:
        id: { type: string }
        displayName: { type: string }
        avatarUrl: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        language: { type: ["string", "null"] }
        currency: { type: ["string", "null"] }
        isModerator: { type: boolean }
        memberSince: { type: string }
        activeListingCount: { type: integer }
        soldListingCount: { type: integer }
    ProfileUpdate:
      type: object
      properties:
        displayName: { type: string, minLength: 1 }
        avatarUrl: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        language: { type: ["string", "null"] }
        currency: { type: ["string", "null"] }
...
    PushTokenInput:
      type: object
      required: [token]
      properties:
        token: { type: string, minLength: 1, maxLength: 400 }
...