# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Sellify API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  currency?: string | null;
  isModerator: boolean;
  memberSince: string;
  activeListingCount?: number;
  soldListingCount?: number;
}